Move OperationVerificationHelper to its own module.
